:recycle: [REFACTOR] 기존 DTO에 대한 의존 제거
필요한 DTO만 남겨두고 나머지 이전 코드 제거
