--- BackEnd/favshare/src/main/java/com/favshare/_temp/dto/PopInFeedDto.java
@@ -1,12 +1,0 @@
-package com.favshare._temp.dto;
-
-import lombok.*;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class PopInFeedDto {
-	private int id;
-	private int popId;
-	private int feedId;
-}
