--- BackEnd/favshare/src/main/java/com/favshare/_temp/dto/PopUserYoutubeDto.java
@@ -1,20 +1,0 @@
-package com.favshare._temp.dto;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class PopUserYoutubeDto {
-	int popId;
-	int userId;
-	String nickname;
-	String userPicture;
-	int youtubeId;
-	String youtubeName;
-	String youtubeUrl;
-	int startSecond;
-	int endSecond;
-	String title;
-
-}
