--- BackEnd/favshare/src/main/java/com/favshare/_temp/dto/input/FeedPopIdDto.java
@@ -1,14 +1,0 @@
-package com.favshare._temp.dto.input;
-
-import java.util.List;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class FeedPopIdDto {
-	int feedId;
-	List<Integer> popId;
-
-}
