--- BackEnd/favshare/src/main/java/com/favshare/_temp/dto/input/IdNameDto.java
@@ -1,12 +1,0 @@
-package com.favshare._temp.dto.input;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class IdNameDto {
-	int Id;
-	String Name;
-
-}
