--- BackEnd/favshare/src/main/java/com/favshare/_temp/dto/input/IdolUserIdDto.java
@@ -1,11 +1,0 @@
-package com.favshare._temp.dto.input;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class IdolUserIdDto {
-	int userId;
-	int idolId;
-}
