--- BackEnd/favshare/src/main/java/com/favshare/_temp/dto/input/UserCommentContentIdDto.java
@@ -1,12 +1,0 @@
-package com.favshare._temp.dto.input;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserCommentContentIdDto {
-	int userId;
-	int commentId;
-	String content;
-}
