--- BackEnd/favshare/src/main/java/com/favshare/_temp/dto/input/UserCommentIdDto.java
@@ -1,12 +1,0 @@
-package com.favshare._temp.dto.input;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserCommentIdDto {
-	int userId;
-	int commentId;
-
-}
