--- BackEnd/favshare/src/main/java/com/favshare/_temp/dto/input/UserPopContentIdDto.java
@@ -1,12 +1,0 @@
-package com.favshare._temp.dto.input;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserPopContentIdDto {
-	int userId;
-	int popId;
-	String content;
-}
