--- BackEnd/favshare/src/main/java/com/favshare/_temp/dto/input/UserPopYoutubeIdDto.java
@@ -1,13 +1,0 @@
-package com.favshare._temp.dto.input;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserPopYoutubeIdDto {
-	int userId;
-	int popId;
-	private int youtubeId;
-
-}
