+++ BackEnd/favshare/src/main/java/com/favshare/follow/dto/FollowDto.java
@@ -1,4 +1,4 @@
-package com.favshare._temp.dto;
+package com.favshare.follow.dto;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
