+++ BackEnd/favshare/src/main/java/com/favshare/follow/dto/FromUserToUserDto.java
@@ -1,4 +1,4 @@
-package com.favshare._temp.dto.input;
+package com.favshare.follow.dto;
 
 import lombok.*;
