+++ BackEnd/favshare/src/main/java/com/favshare/user/controller/UserController.java
@@ -2,8 +2,8 @@
 //
 //import java.util.*;
 //
-//import com.favshare._temp.dto.UserAccountDto;
-//import com.favshare._temp.dto.input.UserSignUpDto;
+//import com.favshare.user.dto.UserAccountDto;
+//import com.favshare.user.dto.UserSignUpDto;
 //import com.favshare.user.service.JwtServiceImpl;
 //import com.favshare.user.service.UserService;
 //import org.springframework.beans.factory.annotation.Autowired;
