+++ BackEnd/favshare/src/main/java/com/favshare/user/controller/UserFeedController.java
@@ -3,9 +3,9 @@ package com.favshare.user.controller;
 import java.util.HashMap;
 
 import com.favshare.pop.dto.pop.PopDto;
-import com.favshare._temp.dto.input.UserPopIdDto;
 import com.favshare.pop.dto.pop.UserProfileDto;
 import com.favshare.pop.service.PopService;
+import com.favshare.user.dto.request.UserPopIdRequest;
 import com.favshare.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -29,10 +29,10 @@ public class UserFeedController {
 
 	@ApiOperation(value = "feed에서 pop 하나 선택 시 pop의 재생 화면 나타내기", response = ResponseEntity.class)
 	@PostMapping("/pop")
-	public ResponseEntity<HashMap<String, Object>> showOnePop(@RequestBody UserPopIdDto userPopIdDto) {
+	public ResponseEntity<HashMap<String, Object>> showOnePop(@RequestBody UserPopIdRequest userPopIdRequest) {
 		HashMap<String, Object> result = new HashMap<String, Object>();
-		int userId = userPopIdDto.getUserId();
-		int popId = userPopIdDto.getPopId();
+		int userId = userPopIdRequest.getUserId();
+		int popId = userPopIdRequest.getPopId();
 
 		try {
 			PopDto popDto = popService.getPopDtoById(userId, popId);
