+++ BackEnd/favshare/src/main/java/com/favshare/user/dto/UserIdMessageDto.java
@@ -1,0 +1,14 @@
+package com.favshare.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserIdMessageDto {
+	int userId;
+	String message;
+
+}
