+++ BackEnd/favshare/src/main/java/com/favshare/user/dto/request/UserInfoDto.java
@@ -1,4 +1,4 @@
-package com.favshare._temp.dto;
+package com.favshare.user.dto.request;
 
 import java.util.Date;
 
@@ -6,33 +6,26 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 import com.favshare.user.entity.User;
 
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @Data
 @NoArgsConstructor
-public class UserAllDto {
+public class UserInfoDto {
 	private int id;
 	private String name;
 	private String email;
 	private String password;
-	private String nickname;
 	@DateTimeFormat(pattern = "yyyy-MM-dd")
 	private Date birthDate;
 	private String phone;
-	private String content;
-	private String profileImageUrl;
 
-	public UserAllDto(User user) {
+	public UserInfoDto(User user) {
 		this.id = user.getId();
 		this.name = user.getName();
 		this.email = user.getEmail();
 		this.password = user.getPassword();
-		this.nickname = user.getNickname();
 		this.birthDate = user.getBirthDate();
 		this.phone = user.getPhone();
-		this.content = user.getContent();
-		this.profileImageUrl = user.getProfileImageUrl();
 	}
 
 }
